requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap-v5==1.0.11
django-cors-headers==3.11.0
django-heroku==0.3.1
django-leaflet==0.28.2
django-registration==3.2
djangorestframework==3.13.1
fontawesomefree==6.1.1
gunicorn==20.1.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
-e git+https://github.com/makinacorpus/django-leaflet.git@bb235e1cb3d9e449cf38f58b75dd6e98082eb799#egg=django_leaflet

